LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/image-src.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <image src="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=3446cc06-91dc-4d63-9d93-8ee6185b2000&amp;encodingcheck=&Gbreve;">


FAIL Speculative parsing, document.write(): image-src Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
